+++ basic/src/tags/components/CreateTagForm.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { createTag } from "wasp/client/operations";
 import { Input } from "../../common/Input";
-import { generateBrightColor } from "../colors";
 import { ColorRadioButtons } from "./ColorRadioButtons";
+import { generateBrightColor } from "./colors";
 import { TagLabel } from "./TagLabel";
 
 interface CreateTagFormProps {
