+++ basic/src/tags/components/colors.ts
